combine msalbase files
