--- internal/msalbase/Account.go
@@ -1,58 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-package msalbase
-
-import (
-	"strings"
-)
-
-//Account represents a user's account with information from an ID token
-type Account struct {
-	HomeAccountID     string `json:"home_account_id,omitempty"`
-	Environment       string `json:"environment,omitempty"`
-	Realm             string `json:"realm,omitempty"`
-	LocalAccountID    string `json:"local_account_id,omitempty"`
-	AuthorityType     string `json:"authority_type,omitempty"`
-	PreferredUsername string `json:"username,omitempty"`
-	GivenName         string `json:"given_name,omitempty"`
-	FamilyName        string `json:"family_name,omitempty"`
-	MiddleName        string `json:"middle_name,omitempty"`
-	Name              string `json:"name,omitempty"`
-	AlternativeID     string `json:"alternative_account_id,omitempty"`
-	RawClientInfo     string `json:"client_info,omitempty"`
-
-	AdditionalFields map[string]interface{}
-}
-
-// NewAccount creates an account.
-func NewAccount(homeAccountID, env, realm, localAccountID, authorityType, username string) Account {
-	return Account{
-		HomeAccountID:     homeAccountID,
-		Environment:       env,
-		Realm:             realm,
-		LocalAccountID:    localAccountID,
-		AuthorityType:     authorityType,
-		PreferredUsername: username,
-	}
-}
-
-// Key creates the key for storing accounts in the cache.
-func (acc Account) Key() string {
-	return strings.Join([]string{acc.HomeAccountID, acc.Environment, acc.Realm}, CacheKeySeparator)
-}
-
-//GetUsername returns the username of an account
-func (acc Account) GetUsername() string {
-	return acc.PreferredUsername
-}
-
-//GetHomeAccountID returns the home account ID of an account
-func (acc Account) GetHomeAccountID() string {
-	return acc.HomeAccountID
-}
-
-//GetEnvironment returns the environment of an account
-func (acc Account) GetEnvironment() string {
-	return acc.Environment
-}
