--- internal/msalbase/access_token_provider.go
@@ -1,10 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-package msalbase
-
-type accessTokenProvider interface {
-	GetSecret() string
-	GetExpiresOn() string
-	GetScopes() string
-}
