--- internal/msalbase/auth_parameters_internal.go
@@ -1,44 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-package msalbase
-
-import "github.com/google/uuid"
-
-// AuthorizationType represents the type of token flow.
-type AuthorizationType int
-
-// These are all the types of token flows
-const (
-	AuthorizationTypeNone                  AuthorizationType = iota
-	AuthorizationTypeUsernamePassword                        = iota
-	AuthorizationTypeWindowsIntegratedAuth                   = iota
-	AuthorizationTypeAuthCode                                = iota
-	AuthorizationTypeInteractive                             = iota
-	AuthorizationTypeClientCredentials                       = iota
-	AuthorizationTypeDeviceCode                              = iota
-	AuthorizationTypeRefreshTokenExchange                    = iota
-)
-
-// AuthParametersInternal represents the parameters used for authorization for token acquisition.
-type AuthParametersInternal struct {
-	AuthorityInfo     AuthorityInfo
-	CorrelationID     string
-	Endpoints         AuthorityEndpoints
-	ClientID          string
-	Redirecturi       string
-	HomeaccountID     string
-	Username          string
-	Password          string
-	Scopes            []string
-	AuthorizationType AuthorizationType
-}
-
-// CreateAuthParametersInternal creates an authorization parameters object.
-func CreateAuthParametersInternal(clientID string, authorityInfo AuthorityInfo) AuthParametersInternal {
-	return AuthParametersInternal{
-		ClientID:      clientID,
-		AuthorityInfo: authorityInfo,
-		CorrelationID: uuid.New().String(),
-	}
-}
