--- internal/msalbase/id_token_test.go
@@ -1,24 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-package msalbase
-
-import (
-	"reflect"
-	"testing"
-)
-
-func TestGetLocalAccountID(t *testing.T) {
-	id := &IDToken{
-		Subject: "sub",
-	}
-	actualLID := id.GetLocalAccountID()
-	if !reflect.DeepEqual("sub", actualLID) {
-		t.Errorf("Expected local account ID sub differs from actual local account ID %s", actualLID)
-	}
-	id.Oid = "oid"
-	actualLID = id.GetLocalAccountID()
-	if !reflect.DeepEqual("oid", actualLID) {
-		t.Errorf("Expected local account ID oid differs from actual local account ID %s", actualLID)
-	}
-}
